# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Business Content API",
        "description": "",
        "version": "1.0.0"
    },
    "paths": {
        "/applications": {
            "get": {
                "tags": [
                    "Application"
                ],
                "summary": "Applications show all or by user id or by startup id",
                "operationId": "6c3dabe293dd0215cc96b29aa1cb9517",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "User id"
                    },
                    {
                        "name": "startup_id",
                        "in": "query",
                        "description": "Startup id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Dataset"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/application": {
            "put": {
                "tags": [
                    "Application"
                ],
                "summary": "Application update",
                "operationId": "223190a7327fb6efe0f1e70818f8ade9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Application"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Application"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Application"
                ],
                "summary": "Create application",
                "operationId": "256aa1676f4f685279dc277be175a27b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Application"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Application"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/application/{id}": {
            "get": {
                "tags": [
                    "Application"
                ],
                "summary": "Application get by ID",
                "operationId": "74b23cfff2f7c78ce98fc756866ff264",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Application id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Application"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Application"
                ],
                "summary": "Delete application",
                "operationId": "b7483738c7780e3e6c02dc230d016809",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Application id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/datasets": {
            "get": {
                "tags": [
                    "Dataset"
                ],
                "summary": "Datasets show all or by user id",
                "operationId": "03a25edb34565c70dbac2d4cbb96f1fd",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "User id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Dataset"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/dataset": {
            "put": {
                "tags": [
                    "Dataset"
                ],
                "summary": "Dataset update",
                "operationId": "9f29a9eb4214f235b0b77f955efa1c3e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Dataset"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Dataset"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Dataset"
                ],
                "summary": "Create dataset",
                "operationId": "f47ff7909b04edaae7dc24030a98d564",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Dataset"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Dataset"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/dataset/{id}": {
            "get": {
                "tags": [
                    "Dataset"
                ],
                "summary": "Get dataset by ID",
                "operationId": "41252879c96db42cd62010f8ce87f4de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Dataset id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Dataset"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Dataset"
                ],
                "summary": "Delete dataset",
                "operationId": "0beaf6bd859a6e67cd33b61d0dba028b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Dataset id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Application": {
                "title": "Application",
                "description": "Application model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "title": "user_id",
                        "description": "user_id",
                        "type": "integer",
                        "example": 1
                    },
                    "startup_id": {
                        "title": "startup_id",
                        "description": "startup_id",
                        "type": "integer",
                        "example": 1
                    },
                    "message": {
                        "title": "message",
                        "description": "message",
                        "type": "string",
                        "example": "I want to work hard and cry about it"
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "created_at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2022-01-22T21:34:30.000000"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "updated_at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2022-01-22T21:34:30.000000"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Application"
                }
            },
            "Dataset": {
                "title": "Dataset",
                "description": "Dataset model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "title": "user_id",
                        "description": "user_id",
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "title": "type",
                        "description": "type",
                        "type": "string",
                        "example": "CREATOR"
                    },
                    "content": {
                        "title": "content",
                        "description": "content",
                        "type": "string",
                        "example": "Documents 123"
                    },
                    "status": {
                        "title": "status",
                        "description": "status",
                        "type": "string",
                        "example": "PENDING"
                    },
                    "comment": {
                        "title": "comment",
                        "description": "comment",
                        "type": "string",
                        "example": "Wrong name"
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "created_at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2022-01-22T21:34:30.000000"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "updated_at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2022-01-22T21:34:30.000000"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Dataset"
                }
            },
...
        }
    },
...
}